vhost: destroy device when all vqs are inactive
We change the behavior of vhost_user_get_vring_base(). Previosly,
destroying a virtqueue will cause the whole device to be destroyed.
The behavior is not specified in the vhost-user protocol.

Thus, we refactor this part. The device will be destroyed only when
all virtqueues in the device are going to be destroyed.

This helps us to simplify the implementation when resetting a virtqueue.

Signed-off-by: Kangjie Xu <kangjie.xu@linux.alibaba.com>
Signed-off-by: Xuan Zhuo <xuanzhuo@linux.alibaba.com>